feat: set up a node project for the blog app backend
Installed dependencies including express, dotenv, bcrypt, mongoose and body-parser. Without the body-parser, the req.body properties could not be destructured as they appeared undefined.

A db for the blog app was set-up and connected to, a reader model as well as reader.routes were set-up and tested to work as expected with insomnia.