feat: Set-up nodemailer with gmail.
Installed handlebars and nodemailer dependency packages, and added handleSendEmail.js and generateOtp.js which housed the logic and setup for automatic sending of email from this backend app and utility function to create OTP code on login for users who enabled 2 factor authentication, respectively. Both files are in the utils folder.

Two additional fields: otp and otpExpires were added to the users schema to aid validation with otp on login.

Finally, '/validate-otp' endpoint was added to the loginRoute to do the logic affirming the user has access to the registered email by providing the correct sent OTP within its validity period of 10 minutes.