feat: implemented a patch endpoint which toggles the active status of newsletter subscribers.
Added a new property 'isActive' which the unsubscribe and subscribe endpoints toggle true and false to unsubscribe or subscribe respectively on user's request.

Updated the email templates that notifies subscribers of successful sub and newsletter to contain the unsubscribe link in their footer.

Finally, updated the post blog endpoint to filter new blog creation notification recipients to only include subscribers with property 'isActive: true' for the newsletter email.

This part sending notifications in the create blog endpoint is broken and yet to figure out why it is broken.